created test class for instance management and added 3 methods in it

+++ lerna-iolaus/src/main/java/edu/asu/lerna/iolaus/service/impl/Neo4jInstanceManager.java
@@ -15,7 +15,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import edu.asu.lerna.iolaus.configuration.neo4j.iml.Neo4jRegistry;
+import edu.asu.lerna.iolaus.configuration.neo4j.impl.Neo4jRegistry;
 import edu.asu.lerna.iolaus.domain.INeo4jInstance;
 import edu.asu.lerna.iolaus.domain.implementation.Neo4jInstance;
 import edu.asu.lerna.iolaus.service.INeo4jInstanceManager;
@@ -39,54 +39,60 @@ public class Neo4jInstanceManager implements INeo4jInstanceManager {
 	 */
 	@Override
 	public String addNeo4jInstance(INeo4jInstance instance) {
-		List<INeo4jInstance> fileList=neo4jRegistry.getfileList();
-		int maxId=0;
-		String port=instance.getPort();
-		String host=instance.getHost();
-		if(!checkConnectivity(port,host)&&instance.isActive()==true?true:false)
-		{
-			return "-2";
-		}
-		if(port.equals("")||host.equals(""))//if port and host is null then return "-1"
-			return "-1";
-		//This loop will search for the maximum id in the registry and will set it to the maxId
-		for(INeo4jInstance file:fileList){
-			if(file.getPort().equals(port)&&file.getHost().equalsIgnoreCase(host)){//if combination of port and host already exists then return "0"
-				return "0";
+		if(instance!=null){
+			List<INeo4jInstance> fileList=neo4jRegistry.getfileList();
+			int maxId=0;
+			String port=instance.getPort();
+			String host=instance.getHost();
+			if(port==null || host==null)
+				return null;
+			if(!checkConnectivity(port,host)&&instance.isActive()==true?true:false)
+			{
+				return "-2";
 			}
-
-			if(Integer.parseInt(file.getId())>maxId){ 
-				maxId=Integer.parseInt(file.getId());
+			if(port.equals("")||host.equals(""))//if port and host is null then return "-1"
+				return "-1";
+			//This loop will search for the maximum id in the registry and will set it to the maxId
+			for(INeo4jInstance file:fileList){
+				if(file.getPort().equals(port)&&file.getHost().equalsIgnoreCase(host)){//if combination of port and host already exists then return "0"
+					return "0";
+				}
+	
+				if(Integer.parseInt(file.getId())>maxId){ 
+					maxId=Integer.parseInt(file.getId());
+				}
 			}
-		}
-		String classPath=Neo4jInstanceManager.class.getProtectionDomain().getCodeSource().getLocation().getPath();//gets the class path
-		BufferedWriter bw=null;
-		try{
-			bw=new BufferedWriter(new FileWriter(URLDecoder.decode(classPath.substring(0,classPath.indexOf("classes")),"UTF-8")+"classes/ConfigurationFiles/Neo4jConfiguration"+(maxId+1)+".txt"));
-			bw.append("id:"+(maxId+1)+"\n"); 
-			String modifiedDescription=instance.getDescription().replaceAll("\n", " ").replace("\r","");
-			bw.append("port:"+instance.getPort()+"\n");
-			bw.append("host:"+instance.getHost()+"\n");
-			bw.append("dbPath:"+instance.getDbPath()+"\n");
-			bw.append("description:"+modifiedDescription+"\n");
-			bw.append("active:"+String.valueOf(instance.isActive()==true?true:false)+"\n");
-			bw.append("userName:"+instance.getUserName());
-			bw.close();
-			instance.setId(String.valueOf(maxId+1));
-			neo4jRegistry.getfileList().add(instance);//add new instance to registry.
-		}catch(IOException exception){
-			if(bw==null){
-				logger.info("Error occured while creating the file");
-			}else{
-				try{
-					bw.close();
-					logger.info("Error occured while writting to the file");
-				}catch(IOException exp){
-					System.out.println("Error occured while closing the file");
-				}	
+			String classPath=Neo4jInstanceManager.class.getProtectionDomain().getCodeSource().getLocation().getPath();//gets the class path
+			BufferedWriter bw=null;
+			try{
+				bw=new BufferedWriter(new FileWriter(URLDecoder.decode(classPath.substring(0,classPath.indexOf("classes")),"UTF-8")+"classes/ConfigurationFiles/Neo4jConfiguration"+(maxId+1)+".txt"));
+				bw.append("id:"+(maxId+1)+"\n"); 
+				String modifiedDescription=instance.getDescription().replaceAll("\n", " ").replace("\r","");
+				bw.append("port:"+instance.getPort()+"\n");
+				bw.append("host:"+instance.getHost()+"\n");
+				bw.append("dbPath:"+instance.getDbPath()+"\n");
+				bw.append("description:"+modifiedDescription+"\n");
+				bw.append("active:"+String.valueOf(instance.isActive()==true?true:false)+"\n");
+				bw.append("userName:"+instance.getUserName());
+				bw.close();
+				instance.setId(String.valueOf(maxId+1));
+				neo4jRegistry.getfileList().add(instance);//add new instance to registry.
+			}catch(IOException exception){
+				if(bw==null){
+					logger.info("Error occured while creating the file");
+				}else{
+					try{
+						bw.close();
+						logger.info("Error occured while writting to the file");
+					}catch(IOException exp){
+						System.out.println("Error occured while closing the file");
+					}	
+				}
 			}
+			return String.valueOf(maxId+1);
+		}else{
+			return null;
 		}
-		return String.valueOf(maxId+1);
 	}
 	
 	/**
@@ -115,16 +121,17 @@ public class Neo4jInstanceManager implements INeo4jInstanceManager {
 	 */
 	@Override
 	public void deleteNeo4jInstance(String instanceId) throws UnsupportedEncodingException {
-
-		List<INeo4jInstance> configFileList=neo4jRegistry.getfileList();
-		for(INeo4jInstance configFile:configFileList){
-			if(configFile.getId().equals(instanceId)){
-				configFileList.remove(configFile);
-				String classPath=Neo4jInstanceManager.class.getProtectionDomain().getCodeSource().getLocation().getPath();
-				String filePath=URLDecoder.decode(classPath.substring(0,classPath.indexOf("classes")),"UTF-8")+"classes/ConfigurationFiles/Neo4jConfiguration"+instanceId+".txt";
-				File file = new File(filePath);
-				file.delete();
-				break;
+		if(instanceId!=null){
+			List<INeo4jInstance> configFileList=neo4jRegistry.getfileList();
+			for(INeo4jInstance configFile:configFileList){
+				if(configFile.getId().equals(instanceId)){
+					configFileList.remove(configFile);
+					String classPath=Neo4jInstanceManager.class.getProtectionDomain().getCodeSource().getLocation().getPath();
+					String filePath=URLDecoder.decode(classPath.substring(0,classPath.indexOf("classes")),"UTF-8")+"classes/ConfigurationFiles/Neo4jConfiguration"+instanceId+".txt";
+					File file = new File(filePath);
+					file.delete();
+					break;
+				}
 			}
 		}
 	}
@@ -160,7 +167,7 @@ public class Neo4jInstanceManager implements INeo4jInstanceManager {
 	 */
 	@Override 
 	public INeo4jInstance getInstance(String id){
-		INeo4jInstance instance=new Neo4jInstance();
+		INeo4jInstance instance=null;
 		for(INeo4jInstance confFile:neo4jRegistry.getfileList()){
 			if(confFile.getId().equals(id)){
 				instance=confFile;
